+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # @project-version@ (@build-time@)
 
-* Added a fix for friendly nameplates Name Only mode not working in raids and dungeons for Patch 10.0.5 [Comment #6840]. For the fix to be applied, you first have to enable Name Only mode in the options under General - Blizzard Settings - Names.
+* Updated Healer Tracker widget to support Preservation Evoker [GH-379].
+* Re-worked Healer Tracker widget to better identify healers by spells they use in Retail and Classic.